common: Move Intel specific code to own file
Prepare for AMD specific parsers and move the Intel code into it's
own file.
