+++ common/meson.build
@@ -26,6 +26,7 @@ uefitoolcommon = static_library('uefitoolcommon',
     'meparser.cpp',
     'fitparser.cpp',
     'ffsparser.cpp',
+    'ffsparser_intel.cpp',
     'ffsreport.cpp',
     'peimage.cpp',
     'treeitem.cpp',
